backend/requirements.txt
autopep8==2.0.2
dj-rest-auth
django-allauth==0.54.0
django-anymail==9.1
django-cors-headers==3.14.0
django-environ==0.10.0
django-extensions==3.2.1
django-filter==23.1
django-haystack==3.2.1
django-phonenumber-field==7.0.2
Django==4.2
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
ipython==8.12.0
phonenumbers==8.13.9
Pillow==9.5.0
psycopg2-binary==2.9.6
redis==4.5.4
Werkzeug==2.2.3